Add a basic test
