+++ packages/server-wallet/.env.test
@@ -1,7 +1,7 @@
 SERVER_DB_HOST=localhost
 SERVER_DB_PORT=5432
 SERVER_DB_NAME=server_wallet_test
-SERVER_DB_USER=root
+SERVER_DB_USER=postgres
 
 HUB_DESTINATION='hub'
