+++ packages/server-wallet/package.json
@@ -27,6 +27,7 @@
     "@types/pg": "7.14.0",
     "@types/pino": "6.0.0",
     "@types/prettier": "1.19.0",
+    "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "2.18.0",
     "@typescript-eslint/parser": "2.18.0",
     "body-parser": "1.19.0",
@@ -41,6 +42,7 @@
     "npm-run-all": "4.1.5",
     "pino-pretty": "4.0.0",
     "prettier": "1.19.1",
+    "supertest": "^4.0.2",
     "ts-jest": "25.0.0",
     "ts-node": "8.9.1",
     "typescript": "3.7.5"
